feat(shop): embed Create Tool form inline in shop ops tab
The "Create Tool" tab previously showed only a button that navigated to a
separate /shop/tools/new page. Embed the tool-creation form directly in the
tab so admins can register a tool without leaving the Shop Operations page.

- Move the form (state, submit handler, inputs) into the create tab panel.
- Remove the now-orphaned "+ New Tool" link on /shop/tools.
- Delete the unused /shop/tools/new page.
