+++ checkin-app/src/app/shop/page.tsx
@@ -1,13 +1,43 @@
 "use client";
 
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
-import { Alert, Button, Card, Center, Container, Group, Loader, Stack, Tabs, Text, Title } from '@mantine/core';
+import { Alert, Button, Card, Center, Container, Group, Loader, Stack, Tabs, Text, TextInput, Title } from '@mantine/core';
 
 export default function ShopOpsPage() {
   const { data: session, status } = useSession();
   const router = useRouter();
+
+  const [newToolName, setNewToolName] = useState("");
+  const [newToolGuide, setNewToolGuide] = useState("");
+  const [saving, setSaving] = useState(false);
+  const [createMessage, setCreateMessage] = useState("");
+
+  const handleCreateTool = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setSaving(true);
+    setCreateMessage("");
+
+    try {
+      const res = await fetch('/api/shop/tools', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ name: newToolName, safetyGuide: newToolGuide })
+      });
+
+      if (res.ok) {
+        setCreateMessage("New tool added successfully!");
+        setNewToolName("");
+        setNewToolGuide("");
+      } else {
+        const data = await res.json();
+        setCreateMessage(data.error || "Failed to create tool.");
+      }
+    } finally {
+      setSaving(false);
+    }
+  };
 
   useEffect(() => {
     if (status === "unauthenticated") {
@@ -69,10 +99,39 @@ export default function ShopOpsPage() {
 
         {isAdmin && (
           <Tabs.Panel value="create" pt="lg">
-            <Stack gap="md" align="flex-start">
-              <Text c="dimmed">Register a new tool definition and safety guide into the database.</Text>
-              <Button onClick={() => router.push('/shop/tools/new')}>Create Tool</Button>
-            </Stack>
+            <Card withBorder radius="md" padding="lg" maw={520}>
+              <Text c="dimmed" mb="lg">
+                Define a new piece of shop equipment to begin tracking safety certifications,
+                authorizing Certifiers, and tracking usage.
+              </Text>
+
+              {createMessage && (
+                <Alert color={createMessage.includes('success') ? 'green' : 'red'} mb="md">{createMessage}</Alert>
+              )}
+
+              <form onSubmit={handleCreateTool}>
+                <Stack>
+                  <TextInput
+                    label="Equipment Name"
+                    required
+                    placeholder="e.g. SawStop Table Saw"
+                    value={newToolName}
+                    onChange={(e) => setNewToolName(e.currentTarget.value)}
+                  />
+                  <TextInput
+                    type="url"
+                    label="Safety Guide URL"
+                    placeholder="https://example.com/safety-manual"
+                    description="Optional link to the required reading or manufacturer manual."
+                    value={newToolGuide}
+                    onChange={(e) => setNewToolGuide(e.currentTarget.value)}
+                  />
+                  <Button type="submit" disabled={saving} loading={saving} mt="sm" style={{ alignSelf: 'flex-start' }}>
+                    Create Tool
+                  </Button>
+                </Stack>
+              </form>
+            </Card>
           </Tabs.Panel>
         )}
